modules/mod-107-gpu-computing/exercise-04-gpu-cluster-management/src/main.py
"""
GPU Cluster Management - Main Entry Point

Manage GPU clusters for ML workloads

Usage:
    python -m src.main [options]

Example:
    python -m src.main --help
"""

import logging
import sys
from pathlib import Path
from typing import Optional
import click
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('app.log'),
        logging.StreamHandler(sys.stdout)
    ]
)

logger = logging.getLogger(__name__)


@click.group()
@click.option('--verbose', '-v', is_flag=True, help='Enable verbose logging')
def cli(verbose: bool):
    """
    GPU Cluster Management

    Manage GPU clusters for ML workloads
    """
    if verbose:
        logging.getLogger().setLevel(logging.DEBUG)
        logger.debug("Verbose logging enabled")


@cli.command()
@click.option('--config', '-c', type=click.Path(exists=True), help='Configuration file')
def run(config: Optional[str]):
    """Run the main application"""
    logger.info("Starting application...")

    try:
        # Main application logic goes here
        logger.info("Application started successfully")

        # TODO: Implement main functionality

    except Exception as e:
        logger.error(f"Application error: {e}", exc_info=True)
        sys.exit(1)


@cli.command()
def validate():
    """Validate configuration and connectivity"""
    logger.info("Running validation checks...")

    checks = [
        ("Configuration", check_config),
        ("Dependencies", check_dependencies),
        ("Connectivity", check_connectivity)
    ]

    failed = []
    for name, check_func in checks:
        try:
            logger.info(f"Checking {name}...")
            check_func()
            logger.info(f"✓ {name} check passed")
        except Exception as e:
            logger.error(f"✗ {name} check failed: {e}")
            failed.append(name)

    if failed:
        logger.error(f"Validation failed for: {', '.join(failed)}")
        sys.exit(1)
    else:
        logger.info("✓ All validation checks passed")


def check_config():
    """Check if configuration is valid"""
    # TODO: Implement configuration validation
    pass


def check_dependencies():
    """Check if all dependencies are available"""
    # TODO: Implement dependency checks
    pass


def check_connectivity():
    """Check connectivity to required services"""
    # TODO: Implement connectivity checks
    pass


if __name__ == '__main__':
    cli()
