modules/mod-107-gpu-computing/exercise-04-gpu-cluster-management/requirements.txt
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
nvidia-ml-py3>=7.352.0
pydantic>=2.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest>=7.4.0
python-dotenv>=1.0.0
ray>=2.7.0
requests>=2.31.0
torch>=2.0.0
